refactor: split dashbot action buttons
