+++ lib/dashbot/core/common/widgets/dashbot_action_buttons/dashbot_generate_language_picker_button.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+
+import '../../../../features/chat/models/chat_models.dart';
+import '../../../../features/chat/viewmodel/chat_viewmodel.dart';
+import '../dashbot_action.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class DashbotGenerateLanguagePicker extends ConsumerWidget
+    with DashbotActionMixin {
+  @override
+  final ChatAction action;
+  const DashbotGenerateLanguagePicker({super.key, required this.action});
+
+  List<String> _extractLanguages(dynamic value) {
+    if (value is List) {
+      return value.whereType<String>().toList();
+    }
+    return const [
+      'JavaScript (fetch)',
+      'Python (requests)',
+      'Dart (http)',
+      'Go (net/http)',
+      'cURL',
+    ];
+  }
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final langs = _extractLanguages(action.value);
+    return Wrap(
+      spacing: 6,
+      runSpacing: 6,
+      children: [
+        for (final l in langs)
+          OutlinedButton(
+            onPressed: () {
+              ref.read(chatViewmodelProvider.notifier).sendMessage(
+                    text: 'Please generate code in $l',
+                    type: ChatMessageType.generateCode,
+                  );
+            },
+            child: Text(l, style: const TextStyle(fontSize: 12)),
+          ),
+      ],
+    );
+  }
+}
